requirements.txt
numpy
scipy
pandas
pyyaml
matplotlib
pycocotools